sex_splitter introduced as its own module, removed sex_split.py
